Updated to Short again
Updated to Short again

+++ app.js
@@ -278,7 +278,7 @@ function showNews (source,id){
                       {
                         "title": "Top Headlines",
                         "value": newsList,
-                        "style": "Long"
+                        "style": "Short"
                       }],
                     "footnote": {
                       "text": "News brought to you by News API",
